drf_course/requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
sqlparse==0.5.3
django-extensions==3.2.3
pillow==10.4.0
tzdata==2024.2django-silk==5.3.2
